reorganize time utils, fix timestamp handling
