+++ nominal/cli/run.py
@@ -4,7 +4,7 @@ from typing import Sequence
 
 import click
 
-from .._utils import _parse_timestamp
+from .._timeutils import _parse_timestamp
 from ._utils import BASE_URL_OPTION, TOKEN_OPTION, get_client
